+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Release Notes
 
 
+## v1.0.1 (2023-10-10)
+
+- npm install react-icons --save.
+- template Admin Dashboard #02.
+- https://github.com/AsmrProg-YT/Dashboard-Designs/tree/master/Admin%20Dashboard%20%2302.
+
 
 ## v1.0.0 (2023-10-09)